requirements.txt
#for editing built in widgets for html
django-widget-tweaks

#for exporting expenses as pdf
reportlab

#for exporting expenses as export
openpyxl

#for API functionality
djangorestframework

#for API authentication
djangorestframework-simplejwt